+++ CHANGELOG.md
@@ -16,6 +16,10 @@
 
 ### Bug Fixes
 
+* Ignore `override` functions in `async_without_await` rule.  
+  [SimplyDanny](https://github.com/SimplyDanny)
+  [#6416](https://github.com/realm/SwiftLint/issues/6416)
+
 * Fix false positive in `unneeded_escaping` rule when an escaping closure is used in
   a nested closure preceded by another closure.  
   [SimplyDanny](https://github.com/SimplyDanny)